refactor messaging cli modules
