+++ src/services/messaging/config.py
@@ -1,0 +1,9 @@
+"""Shared configuration options and constants for messaging services."""
+
+from __future__ import annotations
+
+# Default messaging settings
+DEFAULT_MODE: str = "pyautogui"
+DEFAULT_COORDINATE_MODE: str = "8-agent"
+AGENT_COUNT: int = 8
+CAPTAIN_ID: str = "Agent-4"
